cycle 107: 6-agent grind drain + 3 audit-pass (sec/asset/engine)
Grind (6 todos drained):
- docs-r25 ARCH cross-refs: 3 path-depth + 1 en-dash fix (CRITICAL)
- sec-r11 SDL2_mixer CVE: SECURITY.md +11L advisory section
- compat-r11 _Static_assert audit: 24 added, CAUGHT 3 LOAD-BEARING BUGS
  (unsigned long → uint32_t in fx_blaster_config, songposition, task)
- test-r17 sys.exit antipattern: verified already absent (stale todo)
- engine-r22 allocache baseline: new audit doc verifying cycle-89 numbers
- net-r20 IPv6 scope triage: new triage doc + scope-impl follow-up todo

Audit-pass (3 personas, +9 todos mined):
- security-and-secrets r25 (sentinel a7f2c4d1, 0 HIGH/MED)
- asset-pipeline r26 (sentinel 284eebeb, cycles 96-104 features verified)
- engine-porter r26 (sentinel a5f2d8b7, 5th totalclocklock re-affirmation;
  GAME.C:10129 '//' false-positive resolved — inside /* */ block)

Build: clean. Tests: 1516 passed -m 'not slow' / 3 skipped (~25s).

+++ compat/audio_stub.c
@@ -954,8 +954,8 @@ int MUSIC_PlaySong(unsigned char *song, int loopflag)
 
 void MUSIC_SetContext(int context)                    { music_context = context; }
 int  MUSIC_GetContext(void)                           { return music_context; }
-void MUSIC_SetSongTick(unsigned long t)              { (void)t; }
-void MUSIC_SetSongTime(unsigned long ms)             { (void)ms; }
+void MUSIC_SetSongTick(uint32_t t)              { (void)t; }
+void MUSIC_SetSongTime(uint32_t ms)             { (void)ms; }
 void MUSIC_SetSongPosition(int m, int b, int t)     { (void)m; (void)b; (void)t; }
 
 void MUSIC_GetSongPosition(songposition *pos)
